+++ packages/euclid/src/msgs/vlp.rs
@@ -1,6 +1,5 @@
 use crate::{
     chain::{ChainUid, CrossChainUser},
-    error::ContractError,
     fee::{Fee, TotalFees},
     pool::PoolConfig,
     swap::NextSwapVlp,
@@ -165,22 +164,13 @@ pub struct AllPoolsResponse {
 }
 
 #[cw_serde]
-pub struct MigrateMsg {}
-
-// Function to calculate the asset to be recieved after a swap
-pub fn calculate_swap(
-    swap_amount: Uint128,
-    reserve_in: Uint128,
-    reserve_out: Uint128,
-) -> Result<Uint128, ContractError> {
-    // Calculate the k constant product
-    let k = reserve_in.checked_mul(reserve_out)?;
-    // Calculate the new reserve of token 1
-    let new_reserve_in = reserve_in.checked_add(swap_amount)?;
-    // Calculate the new reserve of token 2
-    let new_reserve_out = k.checked_div(new_reserve_in)?;
-    // Calculate the amount of token 2 to be recieved
-    let token_2_recieved = reserve_out.checked_sub(new_reserve_out)?;
-
-    Ok(token_2_recieved)
+pub struct VlpRemoveLiquidityResponse {
+    pub liquidity_released: PairWithAmount,
+    pub burn_lp_tokens: Uint128,
+    pub tx_id: String,
+    pub sender: CrossChainUser,
+    pub vlp_address: String,
 }
+
+#[cw_serde]
+pub struct MigrateMsg {}
